lib/prompb: use MetricMetadataType type for MetricMetadata.Type field
This eliminates the need of manual conversion between MetricMetadataType and uint32 / int32.

Updates https://github.com/VictoriaMetrics/VictoriaMetrics/issues/2974

This is a follow-up for the commit 5a587f2006bc5aef16a61ca87714817cc21c2116

+++ lib/promscrape/scrapework_test.go
@@ -253,7 +253,7 @@ func testScrapeWorkScrapeInternalSuccess(t *testing.T, streamParse bool) {
 		scrape_timeout_seconds 42 123
 	`, []prompb.MetricMetadata{
 		{
-			Type:             uint32(prompb.MetricMetadataGAUGE),
+			Type:             prompb.MetricMetadataGAUGE,
 			MetricFamilyName: "foo",
 			Help:             "This is test metric.",
 		},
@@ -374,12 +374,12 @@ func testScrapeWorkScrapeInternalSuccess(t *testing.T, streamParse bool) {
 		scrape_timeout_seconds{job="override"} 42 123
 	`, []prompb.MetricMetadata{
 		{
-			Type:             uint32(prompb.MetricMetadataCOUNTER),
+			Type:             prompb.MetricMetadataCOUNTER,
 			MetricFamilyName: "foo",
 			Help:             "This is test metric.",
 		},
 		{
-			Type:             uint32(prompb.MetricMetadataGAUGE),
+			Type:             prompb.MetricMetadataGAUGE,
 			MetricFamilyName: "bar",
 			Help:             "This is another test metric.",
 		},
